Moving the file data to the right place
